requirements.txt
fastapi
pydantic
pydantic-settings
sqlalchemy
pytest
pyodbc
alembic
python-dotenv
uvicorn
playwright
langchain
langchain-openai
langfuse
pymupdf
python-dateutil
pytesseract
PyJWT[crypto]
